project/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.7.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic==2.9.2
